requirements.txt
asgiref==3.4.1
astroid==2.8.2
Django==3.2.8
drf-yasg==1.20.0
isort==5.9.3
lazy-object-proxy==1.6.0
mccabe==0.6.1
platformdirs==2.4.0
pylint==2.11.1
pytz==2021.3
sqlparse==0.4.2
toml==0.10.2
typing-extensions==3.10.0.2
wrapt==1.12.1
djangorestframework==3.12.4
drf-yasg==1.20.0
psycopg2==2.9.1
black==21.9b0
Pillow==8.4.0
djangorestframework-simplejwt==5.0.0